refactor(session): drop the session_ prefix — rename errors/types/server modules
The package path already scopes these names; miles.rollout.session.session_errors
says "session" twice. errors.py, types.py, and server.py follow the package's own
unprefixed precedent (routing-free modules like linear_trajectory.py). Pure
mechanical rename: git mv + import/docstring reference updates, no behavior
change. Class names (SessionServer), CLI args (--session-server-*), and function
names (run_session_server) are unchanged.
